requirements.txt
bitsandbytes==0.45.5
easydict==1.13
flashinfer-python==0.2.2
gradio==5.29.0
langchain==0.3.25
langchain-core==0.3.59
langchain-ollama==0.3.2
langchain-openai==0.3.16
langgraph==0.4.3
langgraph-checkpoint-postgres==2.0.21
lingua-language-detector==2.1.0
langchain_elasticsearch==0.3.2
psycopg==3.2.8
psycopg-binary==3.2.8
psycopg2-binary==2.9.10
psycopg-pool==3.2.6
python-dotenv==1.1.0
vllm==0.8.5post1
transformers==4.53.2
uvicorn==0.34.2
PyPDF2==3.0.1